vscan/uploader.py
import os

URL = 'http://localhost:8000/a/hello'

SCANNER_DIR = '/Users/tyler/code/dimagi/uth-rhd-code/vscan/'

ARCHIVE_PATH = os.path.join(SCANNER_DIR, 'Archive')


class Exam(object):
    def __init__(self, directory, serial, scan_id, date):
        self.serial = serial
        self.scan_id = scan_id
        self.date = date
        self.directory = directory

        self.files = self.collect_files()

    def collect_files(self):
        files = os.listdir(self.directory)
        for f in files:
            if f not in ['.', '..']:
                yield f


def parse_archive():
    exams = os.listdir(ARCHIVE_PATH)
    for exam in exams:
        exam_dir = os.path.join(ARCHIVE_PATH, exam)
        if exam not in ['.', '..'] and os.path.isdir(exam_dir):
            yield Exam(exam_dir, *exam.split('_'))
